# Improve album art performance and compatibility