--- a/controller/apicontroller.php
+++ b/controller/apicontroller.php
@@ -363,26 +363,28 @@ class ApiController extends Controller {
 		if(count($nodes) > 0 ) {
 			// get the first valid node
 			$node = $nodes[0];
-
 			$mime = $node->getMimeType();
-			$content = $node->getContent();
-			return new FileResponse(array('mimetype' => $mime, 'content' => $content));
-		} else {
-			// try to extract the first picture from the first track on the album
-			$tracks = $this->trackBusinessLayer->findAllByAlbum($albumId, $this->userId);
-			$extractor = new \getID3();
-			$metadata = $extractor->analyze('oc://' . $this->userFolder->getById($tracks[0]->getFileId())[0]->getPath());
-			\getid3_lib::CopyTagsToComments($metadata);
 
-			if(array_key_exists("comments", $metadata) &&
-				array_key_exists("picture", $metadata["comments"]) &&
-				!is_null($metadata["comments"]["picture"][0]["data"]) ) {
-				return new FileResponse(array(
-					'mimetype' => $metadata["comments"]["picture"][0]["image_mime"],
-					'content' => $metadata["comments"]["picture"][0]["data"]
-				));
+			if (0 === strpos($mime, 'audio')) { // embedded cover image
+				$extractor = new \getID3();
+				$metadata = $extractor->analyze('oc://' . $node->getPath());
+				\getid3_lib::CopyTagsToComments($metadata);
+
+				if( array_key_exists("comments", $metadata) &&
+					array_key_exists("picture", $metadata["comments"]) &&
+					!is_null($metadata["comments"]["picture"][0]["data"]) ) {
+					return new FileResponse(array(
+							'mimetype' => $metadata["comments"]["picture"][0]["image_mime"],
+							'content' => $metadata["comments"]["picture"][0]["data"]
+					));
+				}
+			}
+			else { // separate image file
+				$content = $node->getContent();
+				return new FileResponse(array('mimetype' => $mime, 'content' => $content));
 			}
 		}
+
 		$r = new Response();
 		$r->setStatus(Http::STATUS_NOT_FOUND);
 		return $r;