--- a/js/bower.json
+++ b/js/bower.json
@@ -9,7 +9,6 @@
     "restangular" : "",
     "soundmanager": "",
     "underscore" : "",
-    "ngQueue" : "",
     "js-cookie" :""
   }
 }