--- a/js/config/app.js
+++ b/js/config/app.js
@@ -33,7 +33,7 @@ if($('html').hasClass('ie')) {
 	setTimeout(replaceSVGs, 5000);
 }
 
-angular.module('Music', ['restangular', 'duScroll', 'gettext', 'ngRoute', 'ngQueue'])
+angular.module('Music', ['restangular', 'duScroll', 'gettext', 'ngRoute'])
 	.config(['RestangularProvider', '$routeProvider',
 		function (RestangularProvider, $routeProvider) {