--- a/js/public/app.js
+++ b/js/public/app.js
@@ -12,7 +12,7 @@ if($('html').hasClass('ie')) {
 	setTimeout(replaceSVGs, 5000);
 }
 
-angular.module('Music', ['restangular', 'duScroll', 'gettext', 'ngRoute', 'ngQueue'])
+angular.module('Music', ['restangular', 'duScroll', 'gettext', 'ngRoute'])
 	.config(['RestangularProvider', '$routeProvider',
 		function (RestangularProvider, $routeProvider) {