--- a/utility/scanner.php
+++ b/utility/scanner.php
@@ -291,16 +291,29 @@ class Scanner extends PublicEmitter {
 			$userId = $this->userId;
 		}
 
-		$remaining = $this->trackBusinessLayer->deleteTrack($fileId, $userId);
+		$result = $this->trackBusinessLayer->deleteTrack($fileId, $userId);
 
-		$this->albumBusinessLayer->deleteById($remaining['albumIds']);
-		$this->artistBusinessLayer->deleteById($remaining['artistIds']);
+		if ($result) { // this was a track file
+			// remove obsolete artists and albums
+			$this->albumBusinessLayer->deleteById($result['obsoleteAlbums']);
+			$this->artistBusinessLayer->deleteById($result['obsoleteArtists']);
 
-		// debug logging
-		$this->logger->log('removed entities - albums: [' . implode(',', $remaining ['albumIds']) .
-			'], artists: [' . implode(',', $remaining['artistIds']) . ']' , 'debug');
+			// check if the removed track was used as embedded cover art file for a remaining album
+			foreach ($result['remainingAlbums'] as $albumId) {
+				$album = $this->albumBusinessLayer->find($albumId, $userId);
+				if ($album->getCoverFileId() == $fileId) {
+					$this->albumBusinessLayer->removeCover($fileId);
+					$this->findEmbeddedCoverForAlbum($albumId, $userId);
+				}
+			}
 
-		$this->albumBusinessLayer->removeCover($fileId);
+			// debug logging
+			$this->logger->log('removed entities - albums: [' . implode(',', $result['obsoleteAlbums']) .
+				'], artists: [' . implode(',', $result['obsoleteArtists']) . ']' , 'debug');
+		}
+		else { // maybe this was an image file
+			$this->albumBusinessLayer->removeCover($fileId);
+		}
 	}
 
 	/**
@@ -457,5 +470,28 @@ class Scanner extends PublicEmitter {
 		}
 
 		return $ordinal;
+	}
+
+	/**
+	 * Loop through the tracks of an album and set the first track containing embedded cover art
+	 * as cover file for the album
+	 * @param int $albumId
+	 * @param int $userId
+	 */
+	private function findEmbeddedCoverForAlbum($albumId, $userId) {
+		if ($this->userFolder != null) {
+			$tracks = $this->trackBusinessLayer->findAllByAlbum($albumId, $userId);
+			foreach ($tracks as $track) {
+				$nodes = $this->userFolder->getById($track->getFileId());
+				if(count($nodes) > 0 ) {
+					// parse the first valid node and check if it contains embedded cover art
+					$image = $this->parseEmbeddedCoverArt($nodes[0]);
+					if ($image != null) {
+						$this->albumBusinessLayer->setCover($track->getFileId(), $albumId);
+						break;
+					}
+				}
+			}
+		}
 	}
 }