… js/vendor/ngQueue/.bower.json
… js/vendor/ngQueue/LICENSE
… js/vendor/ngQueue/README.md
… js/vendor/ngQueue/bower.json
… js/vendor/ngQueue/ngQueue.js