After #517, loading the album cover art was still somewhat slow and did not work at all on IE8 and IE9. It was slow even in case none of the albums in the library had any cover art available; it took time before the placeholders for missing cover art got shown.

With this PR, the loading speed of cover images is improved by an order of magnitude. As a real-life example, my music library contains ~200 albums and almost all of them have embedded cover art set. Before the proposed change, hitting the reload button first took 10s before albums got shown, and then another 30s before all the cover images were loaded. After the change, it still takes 10s before the albums become visible, but then it takes less than 5s to show all the cover images. Placeholders for albums with no cover art are shown immediately.

The slowness and compatibility issues were caused by the fact that the cover images were loaded using javascript to run the HTTP GET queries. When modern browsers are allowed to do the loading directly from URL, they can perform much better. Javascript was used to run the queries because custom handling for failing queries was wanted: cover art was tried to be loaded for all albums and if the query failed, then the placeholder was shown.

The logic was now changed so that the availability of embedded cover art is checked at music scanning time, and the information is saved to the database. The database schema was not changed but the column cover_file_id in the oc_music_albums table is utilized in a new way: if cover for an album is found from a music file, then the file_id of that music file is stored in the column. Previously the column could point only to an image file residing in the same folder with the music tracks. The method ApiController.cover now checks if the pointed file is and image file or an audio file and extracts the image data accordingly.

Note that the music database has to be recreated after this change. If the database is not recreated, only placeholders will be shown for the albums with embedded cover art. It is probably best not to release this change before #518 is fixed, so that the users don't have to recreate the database twice.

Fixes #534.