@scrutinizer-notifier 2016-10-23
A [new inspection](https://scrutinizer-ci.com/g/owncloud/music/inspections/9c7944f8-deff-4ed6-a413-4615e734ee15) was created.