@scrutinizer-notifier 2016-10-23
The [inspection](https://scrutinizer-ci.com/g/owncloud/music/inspections/4bd7025d-95ec-4428-8664-3d5648eaf7d6) completed: **11 updated code elements**