@paulijar 2016-10-30
It should also be noted that this PR changes the app behavior by making the Music app prefer embedded cover images instead of the distinct image files. Previously, embedded cover art was used only as a second choice if no image files were found from the directory of the audio file. 

This change was made to simplify the implementation, but I also think it makes more sense: picture file found from the directory of the audio file may or may not be related to the track in question, but the embedded cover art is definitely related to the track.