@scrutinizer-notifier 2016-11-22
A [new inspection](https://scrutinizer-ci.com/g/owncloud/music/inspections/bdc5b248-e38b-4525-8155-197b70088f37) was created.